+++ app/src/main/java/com/example/androidmvvm/data/api/UserListJson.kt
@@ -1,4 +1,4 @@
-package com.example.androidmvvm.data.user.api
+package com.example.androidmvvm.data.api
 
 import com.squareup.moshi.Json
