+++ app/src/main/java/com/example/androidmvvm/data/db/AppDatabase.kt
@@ -1,9 +1,9 @@
-package com.example.androidmvvm.data
+package com.example.androidmvvm.data.db
 
 import android.arch.persistence.room.Database
 import android.arch.persistence.room.RoomDatabase
-import com.example.androidmvvm.data.user.db.UserDao
-import com.example.androidmvvm.data.user.db.UserEntity
+import com.example.androidmvvm.data.db.UserDao
+import com.example.androidmvvm.data.db.UserEntity
 
 
 @Database(entities = [UserEntity::class], version = 1)
