+++ app/src/main/java/com/example/androidmvvm/data/db/UserEntity.kt
@@ -1,4 +1,4 @@
-package com.example.androidmvvm.data.user.db
+package com.example.androidmvvm.data.db
 
 import android.arch.persistence.room.ColumnInfo
 import android.arch.persistence.room.Entity
