+++ app/src/main/java/com/example/androidmvvm/di/ApplicationComponent.kt
@@ -1,7 +1,7 @@
 package com.example.androidmvvm.di
 
-import com.example.androidmvvm.data.AppDatabase
-import com.example.androidmvvm.data.user.api.UserApi
+import com.example.androidmvvm.data.db.AppDatabase
+import com.example.androidmvvm.data.api.UserApi
 import com.example.androidmvvm.di.user.UserListComponent
 import com.example.androidmvvm.di.user.UserListModule
 import dagger.Component
