+++ app/src/main/java/com/example/androidmvvm/domain/User.kt
@@ -1,4 +1,4 @@
-package com.example.androidmvvm.domain.user
+package com.example.androidmvvm.domain
 
 /**
  * Created by tamayan on 2018/03/27.
